Create project structure suitable for a distributable package
This follows the project structure outlined in
https://pragprog.com/book/bopytest/python-testing-with-pytest
