+++ tests/conftest.py
@@ -1,0 +1,3 @@
+"""Placeholder."""
+
+# nothing here yet
